+++ app/src/main/java/com/nu1lifier/easyweather/Activity_Cities.java
@@ -1,0 +1,36 @@
+package com.nu1lifier.easyweather;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+public class Activity_Cities extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_cities);
+
+        Button backButton = findViewById(R.id.BackButton);
+        Button londonButton = findViewById(R.id.CityLondon);
+
+        londonButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(), Activity_Main.class);
+                startActivity(intent);
+            }
+        });
+
+        backButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(), Activity_Main.class);
+                startActivity(intent);
+            }
+        });
+    }
+}
