+++ app/src/main/java/com/nu1lifier/easyweather/Activity_Settings.java
@@ -2,24 +2,39 @@ package com.nu1lifier.easyweather;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.text.Layout;
 import android.view.View;
+import android.widget.Button;
 import android.widget.ImageView;
 
-public class MainActivity3 extends AppCompatActivity {
+public class Activity_Settings extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_settings);
 
+        Button backButton = findViewById(R.id.BackButton);
         ImageView sun = findViewById(R.id.imageView2);
+
+
+        backButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(), Activity_Main.class);
+                startActivity(intent);
+            }
+        });
+
         sun.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 setTheme(R.style.AppTheme);
             }
         });
+
+
     }
 }
